Change Qualification.when to be a date. (#3)
* Change Qualification.when to be a date.

This is for thewca/worldcubeassociation.org#2051.  Basing a qualification result on a DateTime is more precision than we can really measure -- we don't know at what time during a competition a result was set.  The best we can say for sure is whether the competition start_date or end_date was before the qualification results were due.  I defined this based on start_date to not disincentivize multi-day competitions, though end_date could as easily be used.

In addition:
-change the name of `when` to `whenDate`, in compliance with the style guide
-add support for "type": "ranking", which was present in the Google Doc version of WCIF but not here.

Because Qualification is not implemented currently in the WCA website, I hope that changing the definition is safe.

* Change from attemptResult and ranking to level

* Update specification.md

Co-authored-by: Jonatan Kłosko <jonatanklosko@gmail.com>